Somthing is still wrong with the interaction part, but it is hard to understand what. I will therefor introduce an elliptical trap, like the exercises tell me. Then I can compare directly with Øyvind's results.

+++ Hamiltonians/hamiltonian.h
@@ -13,6 +13,6 @@ public:
 
 protected:
     class System* m_system = nullptr;
-    double m_interactionPotential = 0;
+    int m_interactionPotential = 0;
 };
